+++ backend/routes/auth.js
@@ -63,7 +63,7 @@ router.post('/login',
             return res.status(400).json({ errors: errors.array() });
         }
         try {
-            const { email, password } = req.body;
+            const { email,  } = req.body;
 
             // Checking that if user with the email from req.header present in DataBase
             const user = await User.findOne({ email })
